Move enums to new file
