+++ api/src/utils/enums.ts
@@ -1,0 +1,65 @@
+const interestsEnum = {
+  POLITICS: 'POLITICS',
+  EDUCATION: 'EDUCATION',
+  HOUSING: 'HOUSING',
+  LIT: 'LIT',
+  MUSIC: 'MUSIC',
+  VISUAL_ARTS: 'VISUAL ARTS',
+  STAGE_AND_SCREEN: 'STAGE AND SCREEN',
+  FOOD_AND_LAND: 'FOOD AND LAND',
+  NATURE: 'NATURE',
+  TRANSPORTATION: 'TRANSPORTATION',
+  HEALTH: 'HEALTH',
+  CANNABIS: 'CANNABIS',
+  IMMIGRATION: 'IMMIGRATION',
+  FUN: 'FUN',
+  NONE: 'NONE',
+};
+
+const rolesEnum = {
+  CONTRIBUTOR: 'CONTRIBUTOR',
+  STAFF: 'STAFF',
+  ADMIN: 'ADMIN',
+  TBD: 'TBD',
+};
+
+const racesEnum = {
+  AMERICAN_INDIAN_OR_ALASKAN_NATIVE: 'AMERICAN INDIAN OR ALASKAN NATIVE',
+  BLACK_OR_AFRICAN_AMERICAN: 'BLACK OR AFRICAN AMERICAN',
+  MIDDLE_EASTERN_OR_NORTH_AFRICAN: 'MIDDLE EASTERN OR NORTH AFRICAN',
+  NATIVE_HAWAIIAN_OR_PACIFIC_ISLANDER: 'NATIVE HAWAIIAN OR PACIFIC ISLANDER',
+  LATINX_OR_HISPANIC: 'LATINX OR HISPANIC',
+  WHITE: 'WHITE',
+  ASIAN: 'ASIAN',
+  OTHER: 'OTHER',
+  NONE: 'NONE',
+};
+
+const onboardingStatusEnum = {
+  ONBOARDING_SCHEDULED: 'ONBOARDING_SCHEDULED',
+  STALLED: 'STALLED',
+  ONBOARDED: 'ONBOARDED',
+};
+
+const pitchStatusEnum = {
+  APPROVED: 'APPROVED',
+  PENDING: 'PENDING',
+  REJECTED: 'REJECTED',
+  NONE: 'NONE',
+};
+
+const assignmentStatusEnum = {
+  IN_PROGRESS: 'IN_PROGRESS',
+  COMPLETED: 'COMPLETED',
+  ABANDONED: 'ABANDONED',
+  NONE: 'NONE',
+};
+
+export {
+  interestsEnum,
+  rolesEnum,
+  racesEnum,
+  onboardingStatusEnum,
+  pitchStatusEnum,
+  assignmentStatusEnum,
+};
